.... heh? Major UX update - App feel from medicine to schedule is now similar. - Medicine cardview mirrors the appearance of the Schedule cardview' - Medicine and schedule does not expand on click anymore, instead it opens up a fragment that shows the significant details regarding both. - Adds actual use of AddMedicineActivity   - Adds AddMedicineTypeFragment and AddMedicineDetailsFragment - Changes TabLayout to StepperLayout(Tabs) - Uses callbacks to finish medicine creation - Adds animations and drawables - Adds material-stepper to gradle

+++ app/src/main/java/ph/edu/mobapde/meditake/meditake/beans/Capsule.java
@@ -19,7 +19,7 @@ public class Capsule extends Medicine{
         setCapsuleSpecificInfo();
     }
 
-    public Capsule(String brandName, String genericName, String medicineFor, double amount){
+    public Capsule(String brandName, String genericName, String medicineFor, long amount){
         super(brandName, genericName, medicineFor, amount);
         setCapsuleSpecificInfo();
     }
