+++ app/src/main/java/ph/edu/mobapde/meditake/meditake/util/MedicineUtil.java
@@ -32,7 +32,7 @@ public class MedicineUtil {
         return databaseConnection.createMedicine(medicine);
     }
 
-    public void setMedicineInfo(Medicine medicine, String brandName, String genericName, String medicineFor, double amount, int dosage){
+    public void setMedicineInfo(Medicine medicine, String brandName, String genericName, String medicineFor, long amount, int dosage){
         initializeDBConnection(this.context);
         medicine.setBrandName(brandName);
         medicine.setGenericName(genericName);
